remove handler plugin
