+++ src/plugin/action.rs
@@ -1,4 +1,7 @@
-use dip::cli::SubcommandPlugin;
+use dip::{
+    bevy::app::{App, Plugin},
+    cli::SubcommandPlugin,
+};
 
 mod build;
 mod bundle;
@@ -7,6 +10,14 @@ mod tool;
 pub use build::*;
 pub use bundle::*;
 pub use tool::*;
+
+pub struct DipActionPlugin;
+
+impl Plugin for DipActionPlugin {
+    fn build(&self, app: &mut App) {
+        app.add_plugin(ActionPlugin).add_plugin(BuildActionPlugin);
+    }
+}
 
 #[derive(SubcommandPlugin, clap::Subcommand, Clone, Debug)]
 pub enum Action {
